module4-task2: Декомпозирует Main, Property
