+++ src/components/place-card/place-card.e2e.test.js
@@ -3,7 +3,7 @@ import Enzyme, {mount} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import {PlaceCard} from "./place-card";
 import {cardDataArray, theme} from "../../utils/test.utils";
-import {PlaceCardDetail} from "../place-card-detail/place-card-detail";
+import {Property} from "../property/property";
 import {BrowserRouter as Router} from "react-router-dom";
 
 Enzyme.configure({
@@ -18,7 +18,7 @@ describe(`Place card test`, () => {
             theme={theme}
             offer={cardDataArray[0]}
           >
-            <PlaceCardDetail offers={cardDataArray}/>
+            <Property offers={cardDataArray}/>
           </PlaceCard>
         </Router>
     );
@@ -32,7 +32,7 @@ describe(`Place card test`, () => {
             theme={theme}
             offer={cardDataArray[0]}
           >
-            <PlaceCardDetail offers={cardDataArray}/>
+            <Property offers={cardDataArray}/>
           </PlaceCard>
         </Router>
     );
